Objective-C layer: Fix an erronous assumption about ssize_t being equivalent to int.
darcs-hash:77922acb6a31eeacd3df23d00f9f67f0cd4fa21a

+++ Objective-C/PyObjC/objc-runtime-gnu.m
@@ -10,7 +10,7 @@
 #if defined(GNU_RUNTIME)
 #include "objc-runtime-gnu.h"
 
-struct objc_protocol_list* PyObjCRT_AllocProtocolList(int numProtocols)
+struct objc_protocol_list* PyObjCRT_AllocProtocolList(ssize_t numProtocols)
 {   
 	struct objc_protocol_list *plist;
 	
@@ -33,7 +33,7 @@ int PyObjCRT_SetupClass(
 	const char*name, 
 	Class superCls,
 	Class rootCls,
-	int ivarSize,
+	ssize_t ivarSize,
 	struct objc_ivar_list* 	ivarList,
 	struct objc_protocol_list* protocolList
 )
@@ -128,7 +128,7 @@ void PyObjCRT_ClearClass(Class cls)
 	}
 }
 
-struct objc_method_list *PyObjCRT_AllocMethodList(int numMethods)
+struct objc_method_list *PyObjCRT_AllocMethodList(ssize_t numMethods)
 {
         struct objc_method_list *mlist;
